+++ tests/unit/request/RequestTest.php
@@ -18,10 +18,10 @@
 
 declare(strict_types=1);
 
-namespace gordonmcvey\httpsupport\test\unit;
+namespace gordonmcvey\httpsupport\test\unit\request;
 
 use gordonmcvey\httpsupport\enum\Verbs;
-use gordonmcvey\httpsupport\Request;
+use gordonmcvey\httpsupport\request\Request;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
