+++ tests/unit/response/ResponseStatusHeaderTest.php
@@ -16,13 +16,21 @@
  * limitations under the License.
  */
 
-declare(strict_types=1);
+ declare(strict_types=1);
 
-namespace gordonmcvey\httpsupport;
+ namespace gordonmcvey\httpsupport\test\unit;
+ 
+ use gordonmcvey\httpsupport\enum\statuscodes\SuccessCodes;
+ use gordonmcvey\httpsupport\response\ResponseStatusHeader;
+ use PHPUnit\Framework\Attributes\Test;
+ use PHPUnit\Framework\TestCase;
 
-interface JsonRequestInterface extends RequestInterface
-{
-    public function jsonBody(): mixed;
-
-    public function jsonParam(string $key, mixed $default = null): mixed;
-}
+ class ResponseStatusHeaderTest extends TestCase
+ {
+    #[Test]
+    public function itGeneratesTheExpectedHeaderString(): void
+    {
+        $header = new ResponseStatusHeader(1.1, SuccessCodes::ACCEPTED);
+        $this->assertEquals("HTTP/1.1 202 Accepted", (string) $header);
+    }
+ }
